+++ changelog.md
@@ -286,3 +286,4 @@
 * Fixed advancements and book rewards
 * Improved name distribution
 * Marriage and Birth notifications are now only printed within the village boundaries, or when being friends
+* Added config flag to disable boobs